requirements.txt
PyMuPDF == 1.25.5
python-dotenv == 1.1.0
tqdm == 4.67.1
ipywidgets == 8.1.7
langchain == 0.3.25
langchain-community == 0.3.23
langchain-google-genai == 2.1.4
langchain-huggingface == 0.2.0
langchain-chroma == 0.2.4
sentence-transformers == 4.1.0
transformers == 4.51.3
chromadb == 1.0.9
streamlit == 1.45.1
